Split chat source replacement test support
